requirements.txt
# Core dependencies
torch>=2.0.0
pytorch-lightning>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration and utilities
PyYAML>=6.0
tensorboard>=2.8.0
tqdm>=4.62.0

# Optional: for better performance
numba>=0.56.0

# Development tools (optional)
jupyter>=1.0.0
ipykernel>=6.0.0
